+++ src/main/java/com/realestate/app/domain/property/PropertyDto.java
@@ -1,4 +1,4 @@
-package com.realestate.app.dto;
+package com.realestate.app.domain.property;
 /*
 public record PropertySummaryDto(Long id, String status, double lat, double lng, Long price) {}
 */
